+++ src/test/resources/yelp_5_business_suggestions.json
@@ -1,0 +1,5 @@
+{"unique_id":124,"score":38,"id":124,"localized_completed_text":"ABC Market","location":{"x":-122.459446,"y":37.785371},"country":"US","review_count":2,"review_wilson_score":0.1,"category_aliases":["food","grocery"],"checkin_rate_per_day":0.0,"standardized_score":38,"language_alternate_names":[]}
+{"unique_id":404,"score":32,"id":404,"localized_completed_text":"Philz Coffee","location":{"x":-122.459446,"y":37.785371},"country":"US","review_count":267,"review_wilson_score":0.1,"category_aliases":["coffee","food"],"checkin_rate_per_day":7.1036529680365295,"standardized_score":31233,"language_alternate_names":[]}
+{"unique_id":635,"score":0,"id":635,"localized_completed_text":"Pho Mural","location":{"x":-122.459446,"y":37.785371},"country":"US","review_count":1,"review_wilson_score":0.1,"category_aliases":["localflavor"],"checkin_rate_per_day":0.0,"standardized_score":0,"language_alternate_names":[]}
+{"unique_id":129,"score":67,"id":129,"localized_completed_text":"Antiques and stuff","location":{"x":-122.459446,"y":37.785371},"country":"US","review_count":63,"review_wilson_score":0.1,"category_aliases":["menscloth","fashion","homeandgarden","homedecor","womenscloth","bookstores","shopping","media","vintage","thrift_stores"],"checkin_rate_per_day":0.3194063926940639,"standardized_score":6682,"language_alternate_names":[]}
+{"unique_id":124,"score":29,"id":124,"localized_completed_text":"KT Nails","location":{"x":-122.459446,"y":37.785371},"country":"US","review_count":196,"review_wilson_score":0.1,"category_aliases":["beautysvc","adultentertainment","nightlife","hairremoval","othersalons","waxing"],"checkin_rate_per_day":1.3401826484018264,"standardized_score":130,"language_alternate_names":[]}
